Add files via upload
For noted references, here is the ORIGINAL database for Fitness Unlimited. This organization stands out as a prominent provider of fitness centers, offering a diverse range of fitness programs and membership choices. To effectively manage its operations, Fitness Unlimited maintains a comprehensive retail database to monitor the sales of services and merchandise. In this assignment, I will be customizing the database design to suit the specific requirements of Fitness Unlimited's business model.